Extends icon component
